+++ messages/generic.en
@@ -140,3 +140,15 @@ ADG0000139=Data structure representations is used, use NextData() API
 ADG0000140=Record representations is used, use NextRecord() API
 ADG0000141=Record result set is empty
 ADG0000150=Cannot scan value for type {0}
+ADG0000160=received TCP buffer invalid
+ADG0000161=protocol error, wrong reply received
+ADG0000162=received buffer does not contain data type
+ADG0000163=cluster node evaluation error
+ADG0000164=unexpected buffer type received
+ADG0000165=field length {0} out of range
+ADG0000166=fractional {0} out of range
+ADG0000167=cannot evaluate Fnr
+ADG0000168=end of stream reached
+ADG0000169=less unicode source encoded
+ADG0000170=search string error
+ADG0000171=occurrence index error
